Changed code so full matrix bargraphs can be produced. Since only this graphing capability is being utilized in the current version of the code it is not a problem that no PlotOrder argument is utilized.
Currently the different plot options work, but there is no way to call the same function more than once and pass it different information. If only called once and only listed once in PlotOrder, it should plot each item in list in corresponding row.  If called more than once, it plots the same information but in multiple columns. This should be corrected.